+++ client/gldt_swap_app/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Versions
 
+### [1.1.6] - 2025-07-09
+
+### Fix
+
+- Hero animation not loading
+- GLDT USD price not fetched from API
+
 ### [1.1.5] - 2025-07-03
 
 #### Fix